Crew_assessment/requirements.txt
streamlit
beautifulsoup4
requests
pandas
scikit-learn
python-dotenv
